+++ models/spring-ai-zhipuai/src/main/java/org/springframework/ai/zhipuai/api/ZhiPuAiApi.java
@@ -738,7 +738,7 @@ public class ZhiPuAiApi {
 	@JsonInclude(Include.NON_NULL)
 	public record Embedding(
 			@JsonProperty("index") Integer index,
-			@JsonProperty("embedding") List<Double> embedding,
+			@JsonProperty("embedding") List<Float> embedding,
 			@JsonProperty("object") String object) {
 
 		/**
@@ -747,7 +747,7 @@ public class ZhiPuAiApi {
 		 * @param index The index of the embedding in the list of embeddings.
 		 * @param embedding The embedding vector, which is a list of floats. The length of vector depends on the model.
 		 */
-		public Embedding(Integer index, List<Double> embedding) {
+		public Embedding(Integer index, List<Float> embedding) {
 			this(index, embedding, "embedding");
 		}
 	}
