+++ vector-stores/spring-ai-opensearch-store/src/main/java/org/springframework/ai/vectorstore/OpenSearchVectorStore.java
@@ -150,7 +150,7 @@ public class OpenSearchVectorStore implements VectorStore, InitializingBean {
 				searchRequest.getSimilarityThreshold(), searchRequest.getFilterExpression());
 	}
 
-	public List<Document> similaritySearch(List<Double> embedding, int topK, double similarityThreshold,
+	public List<Document> similaritySearch(List<Float> embedding, int topK, double similarityThreshold,
 			Filter.Expression filterExpression) {
 		return similaritySearch(new org.opensearch.client.opensearch.core.SearchRequest.Builder()
 			.query(getOpenSearchSimilarityQuery(embedding, filterExpression))
@@ -161,7 +161,7 @@ public class OpenSearchVectorStore implements VectorStore, InitializingBean {
 			.build());
 	}
 
-	private Query getOpenSearchSimilarityQuery(List<Double> embedding, Filter.Expression filterExpression) {
+	private Query getOpenSearchSimilarityQuery(List<Float> embedding, Filter.Expression filterExpression) {
 		return Query.of(queryBuilder -> queryBuilder.scriptScore(scriptScoreQueryBuilder -> {
 			scriptScoreQueryBuilder
 				.query(queryBuilder2 -> queryBuilder2.queryString(queryStringQuerybuilder -> queryStringQuerybuilder
